+++ packages/anyhoo_map/lib/anyhoo_map.dart
@@ -5,10 +5,10 @@
 // ignore: unnecessary_library_name
 library anyhoo_map;
 
-export 'src/google_map.dart';
-export 'src/flutter_map.dart';
 export 'src/anyhoo_latlong.dart';
 export 'src/anyhoo_map.dart';
 export 'src/anyhoo_map_type.dart';
 export 'src/anyhoo_marker.dart';
 export 'src/anyhoo_map_settings.dart';
+export 'src/anyhoo_map_controller.dart';
+export 'src/anyhoo_tile_provider.dart';